+++ kroxylicious-api/src/main/java/io/kroxylicious/proxy/filter/package-info.java
@@ -48,7 +48,7 @@
  *   <li>Construction of the filter instance and dispatch of the filter methods {@code onXxxRequest} and
  *       {@code onXxxResponse} takes place on that same thread.</li>
  *   <li>Any computation stages chained to the {@link java.util.concurrent.CompletionStage} returned by
- *       {@link io.kroxylicious.proxy.filter.FilterContext#sendRequest(short, org.apache.kafka.common.protocol.ApiMessage)}
+ *       {@link io.kroxylicious.proxy.filter.FilterContext#sendRequest(org.apache.kafka.common.message.RequestHeaderData, org.apache.kafka.common.protocol.ApiMessage)}
  *       using the default execution methods (using methods without the suffix async) or default asynchronous execution
  *       (using methods with suffix async that employ the stage's default asynchronous execution facility)
  *       are guaranteed to be performed by that same thread.  Computation stages chained using custom asynchronous